Implementa regras de negócio

+++ Academia.Application/Services/EnrollmentService.cs
@@ -1,12 +1,10 @@
 ﻿using Academia.Application.DTOs.Enrollment;
 using Academia.Application.Interfaces;
 using Academia.Domain.Entities;
+using Academia.Domain.Enums;
 using Academia.Domain.Interfaces;
 using Academia.Domain.Validation;
 using AutoMapper;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Academia.Application.Services
 {
@@ -21,13 +19,30 @@ namespace Academia.Application.Services
             _mapper = mapper;
         }
 
-        public async Task<ResponseEnrollment> CreateEnrollment(CreateEnrollment CreateEnrollment)
+        public async Task<ResponseEnrollment> CreateEnrollment(CreateEnrollment createEnrollment)
         {
-            var create = _mapper.Map<Enrollment>(CreateEnrollment);
-            await _unitOfWork.EnrollmentRepository.CreateEnrollment(create);
-            
+            var plan = await _unitOfWork.PlanRepository.GetById(createEnrollment.PlanId);
+            if (plan == null)
+            {
+                throw new NotFoundException("Plan not found");
+            }
+            var student = await _unitOfWork.StudentRepository.GetById(createEnrollment.StudentId);
+            if (student == null)
+            {
+                throw new NotFoundException("Student not found");
+            }
+            if(student.Enrollments.Any(x=>x.EnrollmentSatatus == EnrollmentSatatus.Active))
+            {
+                throw new BadRequestException("Student already registered");
+            }
+
+
+            var create = _mapper.Map<Enrollment>(createEnrollment);
+            var created = await _unitOfWork.EnrollmentRepository.CreateEnrollment(create);
+
+          
             await _unitOfWork.CommitAsync();
-            return _mapper.Map<ResponseEnrollment>(create);
+            return _mapper.Map<ResponseEnrollment>(created);
         }
 
         public async Task<IEnumerable<ResponseEnrollment>> GetAll()
@@ -38,16 +53,21 @@ namespace Academia.Application.Services
 
         public async Task<ResponseEnrollment> GetById(int id)
         {
-            var Enrollment = await _unitOfWork.EnrollmentRepository.GetById(id);
-            if (Enrollment == null)
+            var enrollment = await _unitOfWork.EnrollmentRepository.GetById(id);
+            if(enrollment == null)
             {
-                throw new BadRequestException("Id não encontrado");
+                throw new NotFoundException("Enrollment not found");
             }
-            return _mapper.Map<ResponseEnrollment>(Enrollment);
+            return _mapper.Map<ResponseEnrollment>(enrollment);
         }
 
         public async Task<bool> Remove(int id)
         {
+            var enrollment = await _unitOfWork.EnrollmentRepository.GetById(id);
+            if (enrollment == null)
+            {
+                throw new BadRequestException("Enrollment not found");
+            }
             var result = await _unitOfWork.EnrollmentRepository.Remove(id);
             await _unitOfWork.CommitAsync();
             return result;
