+++ Academia.Application/Services/StudentService.cs
@@ -1,7 +1,9 @@
 ﻿using Academia.Application.DTOs.Student;
 using Academia.Application.Interfaces;
 using Academia.Domain.Entities;
+using Academia.Domain.Enums;
 using Academia.Domain.Interfaces;
+using Academia.Domain.Validation;
 using AutoMapper;
 using System;
 using System.Collections.Generic;
@@ -23,9 +25,15 @@ namespace Academia.Application.Services
         public async Task<ResponseStudent> CreateStudent(CreateStudent createStudent)
         {
             var create = _mapper.Map<Student>(createStudent);
-            await _unitOfWork.StudentRepository.CreateStudent(create);
+            var cpfExist = await _unitOfWork.StudentRepository.CpfExist(create.CPF);
+            if (cpfExist == true)
+            {
+                throw new BadRequestException("Student already registered");
+            }
+
+            var created = await _unitOfWork.StudentRepository.CreateStudent(create);
             await _unitOfWork.CommitAsync();
-            return _mapper.Map<ResponseStudent>(create);
+            return _mapper.Map<ResponseStudent>(created);
         }
 
         public async Task<IEnumerable<ResponseStudent>> GetAll()
@@ -37,11 +45,24 @@ namespace Academia.Application.Services
         public async Task<ResponseStudent> GetById(int id)
         {
             var student = await _unitOfWork.StudentRepository.GetById(id);
+            if(student == null)
+            {
+                throw new NotFoundException("Student not found");
+            }
             return _mapper.Map<ResponseStudent>(student);
         }
 
         public async Task<bool> Remove(int id)
         {
+            var student = await _unitOfWork.StudentRepository.GetById(id);
+            if (student == null)
+            {
+                throw new NotFoundException("Student not found");
+            }
+            if(student.Enrollments.Any(x=>x.EnrollmentSatatus == EnrollmentSatatus.Active))
+            {
+                throw new BadRequestException("It’s not possible to remove a student with an active enrollment");
+            }
             var result = await _unitOfWork.StudentRepository.Remove(id);
             await _unitOfWork.CommitAsync();
             return result;
@@ -49,7 +70,12 @@ namespace Academia.Application.Services
 
         public async Task<ResponseStudent> Update(UpdateDTO update)
         {
+
             var student = await _unitOfWork.StudentRepository.GetById(update.StudentId);
+            if (student == null)
+            {
+                throw new NotFoundException("Student not found");
+            }
             student.AlterarDados(
                  update.Name,
                  update.Email,
